Gitignore and refactoring
Moved snake class to its own file and added gitignore
