+++ .gitignore
@@ -1,0 +1,8 @@
+
+.idea/inspectionProfiles/profiles_settings.xml
+.idea/misc.xml
+.idea/modules.xml
+.idea/snake.iml
+.idea/vcs.xml
+.DS_Store
+.idea/workspace.xml
